+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+
+## [1.12.0] - 2021-03-07
 ### Added
 - A check that we are not on the callback route to prevent infinite looping in case the middleware is added globally
 - Logout method to Guard
@@ -94,7 +96,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Refreshes expired tokens
 - (Cached) resource owner info
 
-[Unreleased]: https://github.com/kronthto/laravel-oauth2-login/compare/v1.11.0...HEAD
+[Unreleased]: https://github.com/kronthto/laravel-oauth2-login/compare/v1.12.0...HEAD
+[1.12.0]: https://github.com/kronthto/laravel-oauth2-login/compare/v1.11.0...v1.12.0
 [1.11.0]: https://github.com/kronthto/laravel-oauth2-login/compare/v1.10.0...v1.11.0
 [1.10.0]: https://github.com/kronthto/laravel-oauth2-login/compare/v1.9.0...v1.10.0
 [1.9.0]: https://github.com/kronthto/laravel-oauth2-login/compare/v1.8.0...v1.9.0